feat(blog): 08-19 批 20 篇草稿全部翻发 + 08-17 修订遗留落库
- 20 篇草稿逐篇过证据门后翻为 production,125 篇全量发布
- 主要实验证据:postgres bloat 44.3% 实测、kafka rebalance 时序
  (pause→PreparingRebalance→消费归零→恢复)、redis 三档吞吐、
  mysql MDL 排队复现、乐观/悲观锁真实 DB 压测、vector ANN 全扫描
- 修复:kafka 镜像 advertised.listeners 覆盖、measure.py 状态解析、
  PG16 round() 兼容、mysql-ddl 自动化复现 (run_all.sh)
- 新增 experiments/mysql-lock-bench 真实 MySQL 锁对比压测
- tests/content.test.ts 更新为 125 个 production slug
- review.md 第 23 节记录本轮证据与边界

+++ experiments/postgres-bloat/03_measure_bloat.sql
@@ -9,7 +9,7 @@
 SELECT relname,
        n_live_tup,
        n_dead_tup,
-       round(100.0 * n_dead_tup / nullif(n_live_tup + n_dead_tup, 0), 1) AS dead_pct_est,
+       round((100.0 * n_dead_tup / nullif(n_live_tup + n_dead_tup, 0))::numeric, 1) AS dead_pct_est,
        last_autovacuum,
        autovacuum_count,
        vacuum_count
@@ -29,8 +29,8 @@ CREATE EXTENSION IF NOT EXISTS pgstattuple;
 SELECT table_len,
        tuple_count,
        dead_tuple_count,
-       round(dead_tuple_percent, 1) AS dead_tuple_percent,
-       round(free_percent, 1) AS free_percent
+       round(dead_tuple_percent::numeric, 1) AS dead_tuple_percent,
+       round(free_percent::numeric, 1) AS free_percent
 FROM pgstattuple('bloat_demo.orders');
 
 -- 4) 顺序扫描计时（膨胀态）：和 04 脚本 VACUUM 后的耗时对比
